CIF-402 - [CommerceTools] As a developer, I want to be able to add multiple payments to the same cart (#34)
Configure a new action called `postCartPayment` to use the same code as `postPayment`.
Update `serverless.yml` definition files
Remove the restriction to only have one payment in the cart.

Update the implementation to support multiple payments for a cart:
* Create another action
* Update the cart mapper to map the mulitple payments in a new field, called `payments`
* Update IT
* Update unit tests

For backwards compatibility the mapper also maps the `payment` field with the first payment in the array. For the same reasons the `CommerceToolsCartPayment` now includes a flag whether it should check that only one payment is allowed in a cart.

Implement the new `deleteCartPayment` function which allows for payments to be removed from the cart.

Add the `methodId` check in integration tests.
Wait for the `delete payment` promise to resolve before returning because the operation may fail.